Excidium rework V1.5 (#1659)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Same exact thing as the previous one:
https://github.com/Rotwood-Vale/Ratwood-Keep/pull/1656 , but now with a
cool new icon and animations! Also some bugfixes
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Same as the previous one. And now with some nice animations and better
looking sprite.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->