Keep That Thang Loaded (#1664)
## About The Pull Request
Gives the Priest a special specifically for spawning with dual
runelocks. "Runic Faith".
Unloaded, of course.
Alongside firearms skill rather high.
Also moves some of their items around to fit it, so nothing needs to be
deleted. They'll need to mug the Inquisitor for pistol shot or go to the
Artificer.
## Why It's Good For The Game
Is this good? NO. It's funny. That's why I propose it.