Bandit Hoardmaster Balancing+ no more skeleton steel. (#1661)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This PR Seeks to correct the issues created by the addition of the
hoardmaster mechanic, by increasing certain upper tier gear hoard
prices, and decreasing the bandits ability to PVE for loot by stripping
skeleton NPC's of steel gear.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
As it stands now, The Hoard master functions by a bandit providing any
offering of value, and that value is added to every single bandit's
personal favor fund. This by itself, means that bandits are getting gear
much, much faster then prior, as 5 bandits means you have 4 other people
adding to the favors that you can use. It also is not helped by the fact
that the hoard master's prices are a little out of wack, with the price
for higher tier stuff being only marginally more then the stuff they
would want to use otherwise. This PR increases prices across the board
for steel grade gear and weaponry.

Further more: Back when the server had a 'dungeon' for adventurers, It
drove Bandits to frequently loot it long before adventurers could ever
form a party to go raid it themselves, leaving them doing PVE for
valuable loot that robbing people just doesn't accomplish. After we
noticed that trend, a PR was made to prevent bandits from entering the
dungeon, so they'd have to fight adventurers for their loot- Which is an
idea circumstance. A similar thing happens currently ever since a PR
that adjusted necromancy/lichs, Skeletons recieved steel gear, including
coifs and gauntlets and swords.

This has lead to bandit players ignoring robbing players, because
they're not really carrying anything of worth anymore comparatively, and
instead just going to pve 'hotspots/dungeons' not even for the dungeons
loot, but specifically to fight large amounts of skeletons and take
their steel/iron. Which results in bandits having maxed out gear in 30
minutes, while interacting with no players. Removing steel equipment
from skeletons lessens this to a dramatic degree, and should result in
less PVE, or less maxed bandits without interacting with others.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->