Remove the underdark (#1663)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->
Removes the underdark and homogenizes the areas in that z-layer.
Adds mushrooms to the bog.

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

The underdark is a huge randomly generated area that is taxing on the
server initialization, is mostly empty and useless and used by no one.