Hunted Flaw
Adds a flaw called 'Hunted'.
Provides a bounty roundstart with a varying value between 250-850, for one of the following crimes:
 - Espionage
 - Sedition
 - Grand Heresy

Found it to be a funny idea. An alternative to something like No-Flaw, since it only gives you a bounty instead of a character defining change like missing limbs. Not that it's a small trade off, considering it's a fairly damn heavy bounty, and would force you to RP it atop that to avoid losing your head.