Cure rot Fix+sawbone treatment to 4 (#1671)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Fixes not being able to cure rot as a pestran or being unable to debride
it as a physician on both player deadites, and rotted corpses.
Also reduces sawbones knife skill to 3 and gives them level 4 treatment
(So they are able to revive)
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Cure rots been broken for like a week due to rotted players no longer
getting the zombie antag status. I removed a check that was causing cure
rot to fail. Simple as.
Also apparently sawbones can't revive people without level 4 treatment,
which is half the point of the role. So.... yeah. That needed fixing.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->