Goblin Touchups | Goblin Shaman (#1667)
## About The Pull Request
New:
- Goblin Shaman introduced. Acts as a healer for Goblins. Locked into
having four spells by default. Prestidigitation, Fetch, Lesser Heal and
Diagnose. They're a mixture of clergy and physicians, for what it's
worth, even spawning with the surgical kits.
- All Goblin roles are given Orcish, if they don't already possess the
language, as a fallback for if staff wants to make non-goblins a Goblin
role. As dumb as that might sound.
 - Goblin Chief is now unlocked to Half Orcs by default.
- Goblin 'Boltslinger' introduced as a Guard class. Just a crossbowmen
of low skill, on par with the other classes.
- Goblins are now in the same faction as NPC goblins, making them
allied.

Changes:
- You're not required to be a Graggarite pre-spawn to play as Goblins.
Instead, it's forced after spawn onto your character, changing whatever
religion you had.
- Goblin travel tiles on the interior can be used by anyone, allowing
escape. The outside tiles are still locked to Goblin trait characters
exclusively.
- Goblins are given 'goblin eyes', which, in theory, should allow night
vision, but they're fucky and don't work well at all in this purpose. If
at all.
- Mineral drops around the Goblin's camp has been significantly
improved, with smatterings of medium and high yield rock all over the
place.
- Goblin Smithery outfitted by default with basic equipment. Two
furnaces, an anvil, a forge and a barrel of water.
- Stone around the encampment modified to hug the walls, so there's no
dead space for mining.
 - Rabble moved into the encampment by default.
- Significantly ups the stats of the Chief, so he can hold his own in a
fight.

## Why It's Good For The Game
Goblins need a lot of QoL stuff. I'm not too happy with how they are
currently, and this rectifies a number of issues, while also introducing
some new content.