Removal of loveless & Pacifism Specials (#1675)
## About The Pull Request
Pacifism was cancerous and genuinely made certain elements of the game,
even outside of combat, actually unplayable. Like, genuinely unable to
interact with stuff. It's better restricted to pacifism bans, as with
Loveless, which is to follow.

Loveless doesn't make sense as a special, because it locks you out of
the following:
 - Receiving / Giving Hugs (Causes a 'freak out', atop that.)
 - Receiving / Giving Kisses (Causes a 'freak out', atop that.)
 - Literally anything involving mechanical ERP.

## Why It's Good For The Game

Pacifism doesn't even need to be argued, does it?

As for Loveless, it was intended as a curse, not a a player special.
Rolling this can ruin some folks' rounds, for better or worse. As funny
as it is. It also makes certain roles requiring mechanics completely
impossible to play, such as Bath Wench.