Maniac Rework-ish (#1673)
## About The Pull Request
After a brief discussion in VC, it was argued by Gyran that Maniac could
make a return, if he wasn't so difficult to kill. At his request, I've
removed most of what made the antag difficult to handle. Traits,
guaranteed 20 to a few stats, etc. I still feel it's incompatible with
the server, as it /requires/ killing and removing folks from the round,
as opposed to more RP based incentives, but hopefully this is a step in
the right direction and should make it far more approachable.

Removes most of the traits that makes maniac impossible to handle
normally.
 - Critical Resistance (Immunity to certain crits. / Harder to crit.)
 - No Stamina
 - No Pain
 - No Pain Stun
 - Hard Dismemberment
 - No Sleep
- Shock Immune (More-or-less kneestinger immunity. Antimagic handles
zaps from lightning.)
 - Stable Heart (Immunity to heart attacks.)
- Stable Liver (Immunity to toxin generation from a damaged / destroyed
liver.)
 - Antimagic (Immunity to magic.)
 - Bloodloss Immunity (Immunity to bloodloss.)

Maniac no longer gets some stats maxed out, instead stuck with what
they've started with:
 - Strength no longer forced to 20.
 - Endurance no longer forced to 20.
 - Constitution no longer forced to 20.

One stat has been adjusted:
 - Unarmed from 4 to 5, putting it on par with their wrestling.

The rest of their stats are untouched, staying at:
 - Knives 6
 - Wrestling 5
 - Treatment 3

References to the following have been removed:
 - NT
 - INRL

## Why It's Good For The Game
Arguably, the antag is incompatible with the server. This is just an
attempt at making it at least somewhat reasonable.