Goblin Camp Exits:
Provides two additional exits to the Goblin camp. One under the mines, and one next to the fishing hut. Images to follow.

They needed additional exits, due to it being too easy to camp one of them at a time. Might be overkill with where I placed them, but we'll see.