Stat Randomisation Amendment (#1687)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Changes the stat randomisation code so that it has an equal chance of
giving you -1, +1, or no stat difference, instead of being skewed to
give you -1 more often than anything else.

## Why It's Good For The Game

Code used to be written by a drooling mongrel. It is now written by a
drooling sexpest mongrel.