Fixes a critical HREF exploit that could be leveraged for ACE. (#1694)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Fixes the ability for staffmembers to have code ran on their clients.

Fixes the ability for the Hoardmaster to spawn any supply pack, leading
to infinite favor glitches.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
CRITICAL BUGFIX
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->