Goblin Camp Exits (#1686)
## About The Pull Request
Provides two additional exits to the Goblin camp. One under the mines,
and one next to the fishing hut. Images to follow.

![image](https://github.com/user-attachments/assets/7bfd9bec-1b06-4f70-90c0-153bb2122f6b)


![image](https://github.com/user-attachments/assets/216f50f7-7484-4d9d-b7aa-977f0aa2a259)

![image](https://github.com/user-attachments/assets/a63e6377-c8de-4eea-82f8-d1e7a395ecdd)

## Why It's Good For The Game
They needed additional exits, due to it being too easy to camp one of
them at a time. Might be overkill with where I placed them, but we'll
see.