Hunted Flaw (#1677)
## About The Pull Request
CHANGES TO MAKE:
 - Have it consider the person's role.
- Maybe fix the machine exploit. Y'know the one. But it's unrelated to
this PR, and has always been an issue.

Adds a flaw called 'Hunted'.
Provides a bounty roundstart with a varying value between 250-850, for
one of the following crimes:
 - Espionage
 - Sedition
 - Grand Heresy

![image](https://github.com/user-attachments/assets/bd4510a1-72e5-4567-8f80-2bb56f078581)

## Why It's Good For The Game
Found it to be a funny idea. An alternative to something like No-Flaw,
since it only gives you a bounty instead of a character defining change
like missing limbs. Not that it's a small trade off, considering it's a
fairly damn heavy bounty, and would force you to RP it atop that to
avoid losing your head.