Fixes the zweihander recipe, cleans up some description text for polearms. (#1696)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Does as it says in the title. The zweihander used to be made of 3 iron,
but was changed for 3 steel for no apparent reason... It's stats
remained the same, which meant it was just a worse greatsword...
This also cleans up some descriptors for polearms, namely, the bardishe,
billhook and halberd.
The halberd used to be described as "An iron halberd" but upon smelting,
it gave us steel, AND requires steel to make... I fixed the description.
(there's 4 commits, because i can't use this site for shit apparently,
dont stake me for it lol)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Swag code cleanup cuz i was bored. Not much is added really.
No one wants to make the "greatsword lite" (zweihander) if they could
have the better option for the same steel cost.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->