Gives firearm skills to experienced guards + Gives watchman skills to sheriff (#1704)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

-------------- Gives firearm skills to experienced guards
---------------
Well this should've probably been done on the firearms PR, but oh well.
Here's the list of the skill given to the guards that should
realistically be trained with firearms, considering their position and
why i think that level of skill makes sense:

- Gatemaster: **Master** If anyone's supposed to use firearms it's this
guy, considering how 90% of his combat is probably going to be ranged,
the gatemaster should be the one most trained to use firearms correctly,
for the same reason he's a master with crossbows and bows too. Because
when he needs to fight, he *will* do it ranged.


- Veteran: **Expert** Much has to be said? He's a retired warrior that
has survived countless battles, it'd make sense for him to remember how
to operate a firearm like if it was second nature.

- Captain: **Expert** The captain is supposed to be the best around,
he's the one that has to train all the men-at-arms and all! So at the
very least he should have an expertise. Doubt it'd be much higher
considering how he still uses mainly weapons like swords.

Knights don't get skilled because of the reason gyran stated below.

- Marshal: **Journeyman** He's not good with many weapons, but if
someone that does mostly paperwork and laws *had* to choose a weapon to
keep with himself, i can assure you it'd be a pistol and decorated
sabre. Why? Expensive and powerful. It's about the same reason the duke
has a firearm. It'd fit him to know how to use one. (THIS DOESN'T GIVE
HIM AN ACTUAL GUN, JUST TRAINING TO BE ABLE TO.)

- Sheriff: **Apprentice** He'd probably be trained with some basic
firearm knowledge after being a watchman for so long, he may not be
exactly proficient with them, but he'll have some idea on how to use one
when it's needed. He's probably seen them used more than any other
watchman around either way.
----------- Gives watchman skills to sheriff --------------

You mean to tell me this guy, has been a watchman for longer than any
other watchmen, knows the streets of this city so well, has fought crime
half of his life, *but* has worse skills than the average watchman?
Seriously? This just gives the sheriff the same weapon skills as
watchmen, and garaunteed expertise on maces, since he's supposed to be a
veteran watchman.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Gives the experienced warriors of the keep the ability to use firearms,
considering how they're supposed to be the most capable around, and
gives the sheriff the same skills as the role directly below him, the
watchmen.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->