Hunted Flaw Rework (#1706)
## About The Pull Request
Before reading the rest of this, consider that this is a flaw. Not a
special you opt into. It's something that can drastically alter your
round, leading to mercenaries seeking you out and taking your head.
We've already got a gameplay loop surrounding it, for players who take
the flaw.

Hunted now takes into account one's occupation, adjusting the crime and
bounty accordingly.

![image](https://github.com/user-attachments/assets/a0bbb029-9982-4ae3-8261-0c72d44d01fb)

For Nobles, a bounty ranging between 450 to 850, with the crimes of:
 - Treasonous Intent
 - Abuse of Power

For Courtiers, a bounty ranging between 400 to 650, with the crimes of:
 - Espionage
 - Inhumen Sympathy

For Garrison, a bounty ranging between 125 to 450, with the crimes of:
 - Sedition
 - Abandonment of Post

For the Church, a bounty ranging between 250 to 500, with the crimes of:
 - Communing with the Inhumen
 - Compelling the Congregation
 - Grand Heresy

For Yeoman, a bounty ranging between 200 to 350, with the crimes of:
 - Hawking Stolen Goods
 - Complicit in Smuggling

For Peasants, a bounty ranging between 120 to 200, with the crimes of:
 - Ignoring Lawful Orders
 - Besmirching Nobility

For Mercenaries, a bounty ranging between 125 to 450, with the crimes
of:
 - Anonymous Request
 - Murder of Nobility
 - Abandonment of Contract

For Youngfolk, a bounty ranging between 150 to 250, with the crimes of:
 - Abuse of Superior
 - Theft from Workplace
 - Abuse of Ducal Trust

For Adventurers / Pilgrims, a bounty ranging between 100 to 325, with
the crimes of:
 - Flight from Authority
 - Smuggling
 - Trafficking
 - Poaching
 - Fleecing
 - Debt

For Goblins, a bounty ranging from 50 to 250, with the crimes of:
 - Torture
 - Murder
 - Heretical Ideals

The fallback, should none of the criteria above be met, is a bounty of
45 to 100, with the crimes of:
 - Poaching
 - Fleecing
 - Debt

## Why It's Good For The Game
It was supposed to go into the other PR, before it was merged.
RAAAAAAA!!!!!!!