Add blindfold to the game (#1681)
## About The Pull Request

Per request of
[this](https://github.com/Rotwood-Vale/Ratwood-Keep/issues/1512) issue I
took the liberty to add to the server the item "blindfold" which can be
used to, as you may guess, blind someone when wearing it. I am not very
well-versed into DM, but I found this easy enough to add and I liked the
idea too.

## Why It's Good For The Game

This PR adds a new item - and taking into consideration the
events/scenes that happen in the game -which I believe will enhance
player experience. The blindfold will make a players screen go entirely
black until it is removed, and it can easily be tied up together with
one piece of cloth from the craft menu.

## Evidences

![Screenshot_2](https://github.com/user-attachments/assets/1b9b8518-e09e-465e-b408-9e85d2d48209)

![Screenshot_3](https://github.com/user-attachments/assets/eeee7019-7bf0-436a-aa58-637c7738453f)

![4a51702c60fb4390a787e0016bcc7d75](https://github.com/user-attachments/assets/911afccd-879a-40be-b6f7-ed5451aacc57)

#### Note
It would be cool to have a proper sprite to it, though it escapes my
.......... skills to actually sprite it. If anyone is willing to make
_at least_ an UI icon for it, I think that would be very neat.