Admin messages for bounties (#1715)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

This pull request adds admin messages for every time that a bounty on
someone is completed! This should hopefully mean that people won't be
killing themselves and delivering their own head for mammon, and that
guild clerks can get punished if they start withdrawing all of the
bounties without reason whatsoever.

Some examples of the messages the admins can encounter:
Lethal bounty completed:

![image](https://github.com/user-attachments/assets/44ca744e-1f6d-4340-afb4-4ccc0a060ccb)
Lethal bounty being completed non-lethally:

![image](https://github.com/user-attachments/assets/9dcb3cba-0d7c-4fb9-84b0-22287aa87415)
Custom bounty being completed:

![image](https://github.com/user-attachments/assets/e6274b90-23e6-4b6b-b2ce-b473c0a586ff)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

With the new hunted flaw giving you so much money and being so popular,
people exploiting it and committing suicide for an easy 100 mammon
roundstart are bount to show up. This just simply gives the admins the
tools to detect these exploiters.

Little reminder that non-lethal bounties **CAN ONLY BE PUT UP BY ACTUAL
PLAYERS, NOT THE FLAWS.** So you can't exactly put up a non-lethal
bounty on yourself and complete it non-lethally without losing money due
to taxes.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->