Allows Mewing (#1716)
## About The Pull Request

This allows people to say mew and mewing again. It's that simple!

## Why It's Good For The Game

We literally have three flavor of cat species plus inn cat. If inn cat
can mew, the rest of the kitties can mew too. It is very silly to have a
filter for mimicking a singular cat sound when we have actual cats in
the game.

## Proof of testing.


https://github.com/user-attachments/assets/9f0e2b1c-f755-4ca3-a91d-855d019514d4