Scrying orb special characters fix (#1722)
## About The Pull Request

fixes #959 by changing the way it handles input name. Most likely it was
not decoding correctly names and during the comparison between a name
and the given input, it didn't handle it correctly. This PR *should* fix
the problem.

## Why It's Good For The Game
Players will now be able to scry both "normal names" and names with
symbols in it (usually ' in the middle or something like that).

## Evidences
Testing with a ' in the name

![image](https://github.com/user-attachments/assets/c3bfe4ae-0182-49a3-8b17-3833dfc2b2fe)

Testing with space in name

![image](https://github.com/user-attachments/assets/948c3743-d7aa-4793-991c-30b440a4156b)

Testing with "normal" name

![image](https://github.com/user-attachments/assets/ebcfdcd4-fcfe-484e-babc-fe4535c5466a)