Gives Malum Worshippers The Boon Of The Forgemaster (#1735)
## About The Pull Request

This reduces the stamina cost of smithing for those whom worship Malum
by effectively 3 str points for the purposes of smithing stamina
calculation only.

For layman's reference. The calc is normally 30max - STRx3 per strike
for applying stamloss from smithing (with a bit of extra calc associated
with dominant hand and minimums). Those worshipping Malum have 21max -
STRx3 per strike. Meaning someone with str7 and worshipping malum will
experience the same very minor stamloss a str10+ character experiences
when smithing.

Note: This _only_ applies to smithing. It does not apply to combat or
anything else.

## Why It's Good For The Game

All the gods should have little minor boons. This is one of them! It's
thematic for Malum- god of the forge, to ease the strain of forging for
even the weakest of his followers. Just a tad. Not too much, but it's
noticeable and even a seelie malumite will benefit a touch- though they
will still grow exhausted quickly..

## Proof Of Testing

![Screenshot 2024-11-30
042722](https://github.com/user-attachments/assets/1073ac88-311e-4875-b4ca-a6c94da5d828)
![Screenshot 2024-11-30
042838](https://github.com/user-attachments/assets/242a49d2-caac-4539-8534-4fc595d8bb0a)