[Port] Eoran stuff from Stonekeep | Acolytes, Templars, Clerics (#1727)
Adds Eora Acolytes/Templars/Clerics to the game and associated spells +
sprites.

Changed Eora's description. Now Eora is the goddess of love, life and
beauty as everything is misleading that Eora and Baotha are one and the
same.

Spells:

Eoran Bloom:

With a long cooldown and casting time, summon a Eoran Bud to help
promote "peace" and pacify people.

Eoras Curse:

Makes the target high and drunk. That's all, yeah.

Robes have alternative version (rightclick on it)

(Credit to SELFHELL for code and monochrome9090 for the sprites)

Feat. SayonaraEcho

Plans: 
1) Remove ability to marry people for druids and give this ability to
eoran acolytes (from here
https://github.com/Darkrp-community/OpenKeep/pull/1038/)
2) Add new sprites for templar's tabards 
3) Review the acolytes/cleric's abilities later (they are too weak)


![4e](https://github.com/user-attachments/assets/3fc6bec9-9c65-4d87-82a2-df14e9e26889)

![im6age](https://github.com/user-attachments/assets/2c03ee12-c7af-419b-b209-0a6a66f3c6a3)