Maniac Buff (#1721)
## About The Pull Request
As well, it gives them dark vision, so they don't have to rely on
torches / lamps and can stalk in the dark.

Returns their strength, endurance and constitution increases, though to
16, not 20.

## Why It's Good For The Game
These are changes based on active testing on the server.
I took a hammer far too strong to Maniac, prior to the aforementioned
active testing, so we've taken to buffing it incrementally on the
server, and this is what we settled on being ok-ish.