gives shophands the normal see prices trait (#1753)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
See title
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Shophands get this discount trait despite being the role meant to man
the counter or run errands meaning they cant even do half of their job
properly ?
it used to be able to see the correct prices but it got changed in this
commit on blackstone for some reason
https://github.com/Blackstone-SS13/BLACKSTONE/commit/6e01bc66c73aec0c1cc52feb8141940fab6cba62
https://github.com/Blackstone-SS13/BLACKSTONE/pull/1083
in this PR and it doesnt even mention that it changed shophand
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->