[Port] Bunch of QoL from Azure Keep and Hearthstone (#1759)
1. "Adds a check to mineral door bump to see if you're carrying a
keychain. Automatically try the door with the keychain in your hand. If
it works, it keeps you in memory as the last bumper, and if you are next
to the door when it closes, it auto-tries to lock with the keys in your
active hand." - PPooch
(https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/965)

2. Fixed: Buckets can FINALLY be entirely filled or entirely poured with
one click. There was a bug that ruins the properly loop of pouring from
buckets. There was still liquid in the bucket even if you tried to pour
it all into the barrel.

3. "Permits beds (and subtypes of beds) to be used without buckling to
them (though obviously you have to be prone anyways).This, by extension,
permits multiple people to sleep on the same (normal) bed without only
one receiving the benefits." - DeltaFire15 You can sleep with your lover
or something like that.
(https://github.com/Tree445/Hearthstone/pull/365)

4. "This change lets you do that with ALL item bundles!!! All you need
to do is take an item bundle and click on a turf with it, and if the
turf has and items that go into the bundle, your character will quickly
gather them all up! Also allows players to cut stakes from bundles of
sticks. Because that's nice to have for people making lots of
palisades!" - Tenryumi
(https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/962)

Except arrows/bolts because of new ammo_holder quiver code (2-3 weeks
ago).


https://github.com/user-attachments/assets/89d27e89-ad10-4bd8-8db9-a22b43925b6a


https://github.com/user-attachments/assets/ed6a3b88-220c-4ea0-a01c-ca222bae85cb