Various Sawbones Corrections - Text (#1765)
## About The Pull Request
Corrects a number of grammatical mistakes that players frequently see.
Adjusts the descriptions of some items to be more appropriate, including
removing alluded self inserts.
The only thing I can't do here, that I'd like to do, is edit the
sprites, because I'm uncertain if those fall under a different license
or not. So, for now, they still reference a snake. Funny that.

In general, it cleans up the sawbones.dm SPELL FILE just a tad. Why is
everything in here?
Not deleting the hidden stuff though. Wooooo. Spooky. Because someone
else did that for me lol

## Why It's Good For The Game
Man, I love this question.