Exploit/game abuse removal (#1739)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Three main things, I'll add more as people bring them up to me as long
as this PR is open:

1. Spitfire spell no longer adds 'stacks' of nonexistant fire, that
trigger when someone is set on fire,
2. Steelmasks in hoardmaster as no longer remotely likely to be cheaper
then the sell price of steel even on low rolls of prices.
3. Bandits can no longer use shylocks. Stop trying to act like a towner
and sell to a stockpile for mammons.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Various things are about that really should be fixed, or had unintended
consequences. This is just a few off the top of my head that got
bragged/shared on the discord recently.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->