Slight Buff to Maniac, Remove blue colour to peoples marked by the wonders (#1749)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
A Slight Buff to Maniac which is removing the knocked down and freaking
out when hit by a hallucination, only for them though.
Remove the blue screen when you are marked by the wonders, instead you
get cursed with hallucinations.
I'll try to find a way to remove the mood effect AND the hallucination
effects for marked peoples when the maniac die when i am not sleep
deprived.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
No comment, tried it in a private server, it work for both things.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->