Puts knife in bag (#1777)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

https://discord.com/channels/1240735983276654733/1258555111005880520/1313568071998505002

![image](https://github.com/user-attachments/assets/a68b9d5d-b41c-4965-8561-bb1f6201f54e)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

---------

Co-authored-by: Gyrandola <pasta.frollagg@gmail.com>