Fix leather gloves' sprite not showing. (#1782)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Changed the inventory sprite name, now the item will be able to call the
on_mob sprites since they had a different name.


![image](https://github.com/user-attachments/assets/bf3ea216-0373-4066-a385-01428279dcdd)

![image](https://github.com/user-attachments/assets/7f00d276-5fb1-421a-a292-bc72938d6d7e)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
fixes #1757
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->