Helmets no longer hide hair if they are in your belt. (#1780)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Makes the proc that hides hair to check if the item is in the belt slot,
doing this it won't remove the hair overlay when is not in the head
slot.


https://github.com/user-attachments/assets/b831bdb3-13c7-4124-902d-91ae1d83b51d


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
It fixes a bug.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->