Milling now gives the correct results (#1786)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Made the mill spawn the `mill_result` instead of the `grind_result`


https://github.com/user-attachments/assets/eba39cb9-5645-4ea0-82d2-3ba6ddccc814


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
fixes #1760
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->